+++ frontend/src/app/directives/email-validator/email-validator.directive.spec.ts
@@ -1,0 +1,8 @@
+import { EmailValidatorDirective } from './email-validator.directive';
+
+describe('EmailValidatorDirective', () => {
+  it('should create an instance', () => {
+    const directive = new EmailValidatorDirective();
+    expect(directive).toBeTruthy();
+  });
+});
