+++ frontend/src/app/directives/personal-number-validator/personal-number-validator.directive.spec.ts
@@ -1,0 +1,8 @@
+import { PersonalNumberValidatorDirective } from './personal-number-validator.directive';
+
+describe('PersonalNumberValidatorDirective', () => {
+  it('should create an instance', () => {
+    const directive = new PersonalNumberValidatorDirective();
+    expect(directive).toBeTruthy();
+  });
+});
