+++ frontend/src/app/directives/phone-number-validator/phone-number-validator.directive.spec.ts
@@ -1,0 +1,8 @@
+import { PhoneNumberValidatorDirective } from './phone-number-validator.directive';
+
+describe('PhoneNumberValidatorDirective', () => {
+  it('should create an instance', () => {
+    const directive = new PhoneNumberValidatorDirective();
+    expect(directive).toBeTruthy();
+  });
+});
