+++ frontend/src/app/pages/admin/admin.component.html
@@ -35,9 +35,13 @@
     <mat-card>
       <mat-card-content>
         <form (ngSubmit)="onSubmit()">
-          <mat-form-field class="">
+          <mat-form-field>
             <mat-label>Rodné číslo očkovaného</mat-label>
-            <input matInput [formControl]="personalNumber" placeholder="Zadejte rodné číslo" required>
+            <input matInput [(ngModel)]="personalNumber"
+                   placeholder="Zadejte rodné číslo"
+                   name="personalNumber"
+                   #personalNumber="ngModel"
+                   required personalNumberValidator>
             <mat-hint>Rodné číslo ve formátu RRMMDD/ČČČ(Č)</mat-hint>
           </mat-form-field>
           <button color="primary"
