+++ frontend/src/app/pages/home/home.component.ts
@@ -1,10 +1,9 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { Component, OnDestroy, ViewChild } from '@angular/core';
+import { FormBuilder, NgForm } from '@angular/forms';
 import { PatientEditable } from '@app/model/PatientEditable';
 import { InsuranceCompany } from '@app/model/InsuranceCompany';
 import { QuestionService } from '@app/services/question/question.service';
 import { PatientService } from '@app/services/patient/patient.service';
-import { validateEmail, validatePersonalNumber, validatePhoneNumber } from '@app/validators/form.validators';
 import { AlertService } from '@app/services/alert/alert.service';
 import { ActivatedRoute } from '@angular/router';
 import { PatientData } from '@app/model/PatientData';
@@ -19,11 +18,11 @@ import { ReCaptchaV3Service } from 'ng-recaptcha';
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.scss']
 })
-export class HomeComponent implements OnInit, OnDestroy {
+export class HomeComponent implements OnDestroy {
 
   private _questionsSubscription?: Subscription;
 
-  public patientForm?: FormGroup;
+  @ViewChild('patientForm') patientForm?: NgForm;
 
   public patient: PatientEditable = new PatientEditable();
   public questions: Question[] = [];
@@ -46,17 +45,6 @@ export class HomeComponent implements OnInit, OnDestroy {
 
     this._questionsSubscription = this._questionService.questionsObservable
     .subscribe(questions => this.questions = questions);
-  }
-
-  ngOnInit() {
-    this.patientForm = this._formBuilder.group({
-      firstName: ['', Validators.required],
-      lastName: ['', Validators.required],
-      personalNumber: [this.patient.personalNumber ?? '', [Validators.required, validatePersonalNumber]],
-      insuranceCompany: ['', Validators.required],
-      phoneNumber: ['', [Validators.required, validatePhoneNumber]],
-      email: ['', [Validators.required, validateEmail]]
-    });
   }
 
   ngOnDestroy() {
